Fix escape (#76)
* Fix escape


* Fix escape


---------

+++ cmd/bigquery.go
@@ -27,7 +27,7 @@ func NewBigQueryCmd() *cobra.Command {
 		cdcHost               string
 		cdcPort               int
 		cdcFlushInterval      time.Duration
-		cdcFileSize           int64
+		cdcFileSize           int
 		logFile               string
 		logLevel              string
 
@@ -134,7 +134,7 @@ func NewBigQueryCmd() *cobra.Command {
 	cmd.Flags().StringVar(&cdcHost, "cdc.host", "127.0.0.1", "TiCDC server host")
 	cmd.Flags().IntVar(&cdcPort, "cdc.port", 8300, "TiCDC server port")
 	cmd.Flags().DurationVar(&cdcFlushInterval, "cdc.flush-interval", 60*time.Second, "")
-	cmd.Flags().Int64Var(&cdcFileSize, "cdc.file-size", 64*1024*1024, "")
+	cmd.Flags().IntVar(&cdcFileSize, "cdc.file-size", 64*1024*1024, "")
 	cmd.Flags().StringVar(&logFile, "log.file", "", "log file path")
 	cmd.Flags().StringVar(&logLevel, "log.level", "info", "log level")
